svelte(phase 3): absorb chat-composer-runner into ChatComposer
Move the 1121-LOC runChatComposer verbatim into ChatComposer.svelte's
<script module> (exported, full DI signature preserved); the instance onMount
calls it exactly as before. The model/thinking/slash/mention selectors + chat-api
stay as injected helper modules. Module-scope icon/t feed both the runtime and
the markup.

chat-composer-runner.js is deleted; its 18-case behavioural suite moves to
ChatComposer.test.js with only the import path repointed to the module-script
export — no test rewrites, all 18 still green.
